+++ aldrin/CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `low_level::EventListener::client()`.
 - Add `BusListener::new()`.
 - Add `Discoverer::client()`.
+- Add `Lifetime::client()`.
 
 ### Fixed